Path: Removed PathPrototype
